[TCcnt] Fixed warnings from size_t TCcountExprs and refactored 2 funcs
Where SAC input code(PMvar, COconst2Int,etc.) is needed for size_t args,
used DBUG assertions and left warnings in place to trace back later
when new data type gets introduced.

Refactored TClookupIdsNode again and LFUindexOfMemberIds to use bool
as both are searching for whether a node is a member of a chain, and
it is used as a true or false for the node being a member.
Added/Changed comment to explain the change.

+++ src/libsac2c/arrayopt/algebraic_wlfi.h
@@ -27,7 +27,7 @@ extern int AWLFIfindPrfParent2 (node *arg_node, node *withidids, node **infowith
 
 extern node *AWLFIflattenExpression (node *arg_node, node **vardecs, node **preassigns,
                                      ntype *ztype);
-extern node *AWLFItakeDropIv (int takect, int dropct, node *arg_node, node **vardecs,
+extern node *AWLFItakeDropIv (int takect, size_t dropct, node *arg_node, node **vardecs,
                               node **preassigns);
 extern bool AWLFIcheckProducerWLFoldable (node *arg_node);
 extern bool AWLFIcheckBothFoldable (node *pwlid, node *cwlids, int cwllevel);
